refactor: 봉사 신청 여부 조회 응답 dto 패키지를 변경한다.
